\name{example}
\alias{example.BOC.93.4.paper.tests}
\alias{example.convert.eg.data}
\alias{example.get.eg.raw.data}
\alias{example.like.sub.sample}
\alias{example.like.total.sample}
\alias{example.show.ytoy.cpi}
\alias{example.tests}
\alias{example.truncate.data}
\alias{example.VAR.SVD}
\alias{example.verify.data}
\alias{example.verify.raw.data}

\title{TS Examples}
\description{DSE Package Examples}
\usage{
	example.BOC.93.4.paper.tests()
	example.convert.eg.data()
	example.get.eg.raw.data()
	example.like.sub.sample()
	example.like.total.sample()
	example.show.ytoy.cpi()
	example.tests()
	example.truncate.data()
	example.VAR.SVD()
	example.verify.data()
	example.verify.raw.data()
}

\details{
	These functions run various examples. Their purpose is mainly to 
	illustrate the use of other functions. Print out the functions and
	examine them.
}

\keyword{DSE}
\keyword{ts}

