\name{get.RNG}
\alias{get.RNG}
\alias{get.RNG.default}

\title{get the RND and seed from an object}
\description{
Get the random number generator and seed used to generate an object.
}
\usage{
	get.RNG(obj)}
\arguments{
	\item{obj}{An object generated by simulation.}
}
\value{The random seed used to generate the object. }
\details{Some information about the version is also verified.}
\seealso{
\code{\link{.Random.seed}}
}
\examples{
if(is.R()) data("eg1.DSE.data.diff", package="dse1")
model <- est.VARX.ls(eg1.DSE.data.diff)
sim <- simulate(model)
get.RNG(sim)
}
\keyword{DSE}
\keyword{ts}

