\name{periods.TSdata}
\alias{periods.TSdata}
\alias{periods.TSestModel}
\alias{start.TSdata}
\alias{start.TSestModel}
\alias{end.TSdata}
\alias{end.TSestModel}
\alias{frequency.TSdata}
\alias{frequency.TSestModel}

\title{Specific Methods for tframed Data}
\description{See the generic function description.}
\usage{
    \method{periods}{TSdata}(x)
    \method{periods}{TSestModel}(x)
    \method{start}{TSdata}(x)
    \method{start}{TSestModel}(x)
    \method{end}{TSdata}(x)
    \method{end}{TSestModel}(x)
    \method{frequency}{TSdata}(x)
    \method{frequency}{TSestModel}(x)
}
\arguments{
    \item{x}{a time series object.}
}

\seealso{
    \code{\link{periods}}, 
    \code{\link{start}},
    \code{\link{end}},
    \code{\link{frequency}}
}
%\keyword{DSE}
\keyword{ts}

