\name{series.names.TSdata}
\alias{series.names.TSdata}
\alias{series.names.TSestModel}
\alias{series.names<-.TSdata}
\alias{series.names<-.TSestModel}
\alias{series.names<-.TSmodel}

\title{Series Names Specific Methods}
\description{See the generic function description.}
\usage{
	series.names(...)
	series.names(...) <- names
}

\seealso{ \code{\link{series.names}} }

\keyword{DSE}
\keyword{ts}

