\name{seriesNames.TSdata}
\alias{seriesNames.TSdata}
\alias{seriesNames.TSmodel}
\alias{seriesNames.TSestModel}
\alias{seriesNames<-.TSdata}
\alias{seriesNames<-.TSestModel}
\alias{seriesNames<-.TSmodel}

\title{Series Names Specific Methods}
\description{See the generic function description.}
\usage{
    \method{seriesNames}{TSdata}(x)
    \method{seriesNames}{TSmodel}(x)
    \method{seriesNames}{TSestModel}(x)
    
    seriesNames(data) <- names
}
\arguments{
    \item{x}{an object from which series names can be extracted.}
}

\seealso{ \code{\link{seriesNames}} }

%\keyword{DSE}
\keyword{ts}

