\name{source.info.TSdata}
\alias{source.info.TSdata}
\alias{source.info.TSestModel}
\alias{sourcedb.TSdata}
\alias{sourcedb.TSestModel}
\alias{sourceserver.TSdata}
\alias{sourceserver.TSestModel}
\alias{identifiers.TSdata}
\alias{identifiers.TSestModel}

\title{tfPADI Specific Methods}
\description{See the generic function description.}
\usage{
    \method{source.info}{TSdata}(obj)
    \method{source.info}{TSestModel}(obj)
    
    \method{sourcedb}{TSdata}(obj)
    \method{sourcedb}{TSestModel}(obj)
    
    \method{sourceserver}{TSdata}(obj)
    \method{sourceserver}{TSestModel}(obj)
    
    \method{identifiers}{TSdata}(obj)
    \method{identifiers}{TSestModel}(obj)    
}

\seealso{ \code{\link[dsepadi]{source.info}} }

%\keyword{DSE}
\keyword{ts}

