\name{summary.TSdata}
\alias{summary.TSdata}
\alias{summary.SS}
\alias{summary.ARMA}
\alias{summary.TSestModel}

\alias{print.summary.TSdata}
\alias{print.summary.SS}
\alias{print.summary.ARMA}
\alias{print.summary.TSestModel}

\title{Specific Methods for Summary}
\description{See the generic function description.}
\usage{
	summary(obj)}

\seealso{
	\code{\link{print}}, 
	\code{\link{summary}}
}

\keyword{DSE}
\keyword{ts}

