\name{test.equal.ARMA}
\alias{test.equal.ARMA}
\alias{test.equal.SS}
\alias{test.equal.TSdata}
\alias{test.equal.TSmodel}
\alias{test.equal.TSestModel}

\title{Specific Methods for Testing Equality}
\description{See the generic function description.}
\usage{
	test.equal(obj1, obj2)}

\seealso{ \code{\link{test.equal}} }

\keyword{DSE}
\keyword{ts}

