
\name{dse1-package}
\alias{dse1-package}

\docType{package}

\title{Transition Package from dse Bundle to dse Package.}

\description{Package to add transition from dse Bundle to dse package.
}

\section{Usage}{
   library("dse1") 
}

\details{
The only purpose of this package is to prevent breaking older code that loads 
package dse1 from the old dse bundle. This package dse1 simple requires package
dse.
}

\author{
Paul Gilbert <pgilbert@bank-banque-canada.ca>

Maintainer: Paul Gilbert <pgilbert@bank-banque-canada.ca>  
}
