\name{periods.TSdata}
\alias{periods.TSdata}
\alias{periods.TSestModel}
\alias{start.TSdata}
\alias{end.TSdata}
\alias{frequency.TSdata}
\alias{tframed.TSdata}
\alias{tfwindow.TSdata}
\alias{tbind.TSdata}
\alias{trim.na.TSdata}

\alias{window.TSdata}

\title{Specific Methods for Tframed data}
\description{See the generic function description.}
\usage{
	periods(x)
	start(x)
	end(x)
	frequency(x)
	tframed(x)
	tfwindow(x)
	tbind(x)
	trim.na(x)
}

\seealso{
	\code{\link{periods}}, 
	\code{\link{start}},
	\code{\link{end}},
	\code{\link{frequency}},
	\code{\link{tframed}},
	\code{\link{tfwindow}},
	\code{\link{tbind}},
	\code{\link{trim.na}}
}

\keyword{DSE}
\keyword{ts}

