\name{series.names.TSdata}
\alias{series.names.TSdata}
\alias{series.names.TSestModel}
\alias{series.names<-.TSdata}
\alias{series.names<-.TSestModel}
\alias{series.names<-.TSmodel}

\title{Series Names Specific Methods}
\description{See the generic function description.}
\usage{
    \method{series.names}{TSdata}(x)
    \method{series.names}{TSestModel}(x)
    
    series.names(data) <- names
}

\seealso{ \code{\link{series.names}} }

%\keyword{DSE}
\keyword{ts}

