\name{source.info.TSdata}
\alias{source.info.TSdata}
\alias{source.info.TSestModel}
\alias{sourcedb.TSdata}
\alias{sourcedb.TSestModel}
\alias{sourceserver.TSdata}
\alias{sourceserver.TSestModel}
\alias{identifiers.TSdata}
\alias{identifiers.TSestModel}

\title{tfPADI Specific Methods}
\description{See the generic function description.}
\usage{
	source.info(x)
	sourcedb(x)
	sourceserver(x)
	identifiers(x)
}

\seealso{ \code{\link[dsepadi]{source.info}} }

\keyword{DSE}
\keyword{ts}

