\name{test.equal.ARMA}
\alias{test.equal.ARMA}
\alias{test.equal.SS}
\alias{test.equal.TSdata}
\alias{test.equal.TSmodel}
\alias{test.equal.TSestModel}

\title{Specific Methods for Testing Equality}
\description{See the generic function description.}
\usage{
    \method{test.equal}{ARMA}(obj1, obj2, fuzz=0)
    \method{test.equal}{SS}(obj1, obj2, fuzz=0)
    \method{test.equal}{TSdata}(obj1, obj2, fuzz=1e-16)
    \method{test.equal}{TSmodel}(obj1, obj2, fuzz=0)
    \method{test.equal}{TSestModel}(obj1, obj2, fuzz=0)
    }
\arguments{
    \item{obj1}{see generic method.}
    \item{obj2}{see generic method.}
    \item{fuzz}{see generic method.}
}

\seealso{ \code{\link{test.equal}} }

%\keyword{DSE}
\keyword{ts}

